rename renderers->generators, ModelCodeGenerator -> WarthogModelBuilder
