+++ query-node/substrate-query-framework/cli/src/generators/ModelRenderer.ts
