+++ query-node/substrate-query-framework/cli/src/generators/utils.ts
