+++ query-node/substrate-query-framework/cli/test/helpers/WarthogModelBuilder.test.ts
@@ -1,10 +1,10 @@
-import { DatabaseModelCodeGenerator } from './../../src/helpers/ModelCodeGenerator';
+import { WarthogModelBuilder } from './../../src/helpers/WarthogModelBuilder';
 import { expect } from 'chai';
 
-describe('ModelCodeGenerator', () => {
+describe('WarthogModelBuild', () => {
     it('should add multi-field multi-entity FTSQuery to the model', () => {
-        const generator = new DatabaseModelCodeGenerator('test/fixtures/multiple-entities.graphql');
-        const model = generator.generateWarthogModel()
+        const generator = new WarthogModelBuilder('test/fixtures/multiple-entities.graphql');
+        const model = generator.buildWarthogModel()
 
         expect(model.ftsQueries).length(1, "Should detect a query");
         const query = model.ftsQueries[0];
@@ -21,8 +21,8 @@ describe('ModelCodeGenerator', () => {
     });
 
     it('should detect multiple queries', () => {
-        const generator = new DatabaseModelCodeGenerator('test/fixtures/multiple-queries.graphql');
-        const model = generator.generateWarthogModel();
+        const generator = new WarthogModelBuilder('test/fixtures/multiple-queries.graphql');
+        const model = generator.buildWarthogModel();
         expect(model.ftsQueries).length(2, "Should detect two queries");
     })
 });
